# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Telegram Multi-Account Bot Manager API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type AccountStatus = typeof AccountStatus[keyof typeof AccountStatus];


export const AccountStatus = {
  active: 'active',
  paused: 'paused',
  banned: 'banned',
  flood_wait: 'flood_wait',
  channels_limit: 'channels_limit',
} as const;

export interface Account {
  id: number;
  phone: string;
  /** @nullable */
  label?: string | null;
  status: AccountStatus;
  joinedCount: number;
  failedCount: number;
  dailyLimit: number;
  /** Current delay in seconds before next join */
  currentDelay: number;
  /**
     * ISO timestamp when flood wait expires
     * @nullable
     */
  floodWaitUntil?: string | null;
  /** @nullable */
  sessionFile?: string | null;
  isPremium?: boolean;
  createdAt: string;
}

export interface AccountInput {
  phone: string;
  label?: string;
  sessionFile?: string;
  isPremium?: boolean;
  dailyLimit?: number;
}

export type AccountUpdateStatus = typeof AccountUpdateStatus[keyof typeof AccountUpdateStatus];


export const AccountUpdateStatus = {
  active: 'active',
  paused: 'paused',
  banned: 'banned',
  flood_wait: 'flood_wait',
  channels_limit: 'channels_limit',
} as const;

export interface AccountUpdate {
  label?: string;
  status?: AccountUpdateStatus;
  dailyLimit?: number;
  isPremium?: boolean;
}

export interface AccountsStats {
  total: number;
  active: number;
  paused: number;
  banned: number;
  floodWait: number;
  totalJoined: number;
  totalFailed: number;
}

export type GroupLinkStatus = typeof GroupLinkStatus[keyof typeof GroupLinkStatus];


export const GroupLinkStatus = {
  pending: 'pending',
  joined: 'joined',
  failed: 'failed',
  skipped: 'skipped',
} as const;

/**
 * @nullable
 */
export type GroupLinkGroupType = typeof GroupLinkGroupType[keyof typeof GroupLinkGroupType] | null;


export const GroupLinkGroupType = {
  group: 'group',
  channel: 'channel',
} as const;

export interface GroupLink {
  id: number;
  url: string;
  status: GroupLinkStatus;
  /** @nullable */
  failReason?: string | null;
  /** @nullable */
  groupTitle?: string | null;
  /** @nullable */
  groupType?: GroupLinkGroupType;
  /**
     * Collection name this came from
     * @nullable
     */
  source?: string | null;
  /** @nullable */
  usedByAccountId?: number | null;
  createdAt: string;
  /** @nullable */
  processedAt?: string | null;
}

export interface LinkInput {
  url: string;
  source?: string;
}

export interface BulkLinksInput {
  urls: string[];
  source?: string;
}

export interface BulkAddResult {
  added: number;
  duplicates: number;
  total: number;
}

export interface LinksStats {
  total: number;
  pending: number;
  joined: number;
  failed: number;
  skipped: number;
}

export type JoinJobStatus = typeof JoinJobStatus[keyof typeof JoinJobStatus];


export const JoinJobStatus = {
  success: 'success',
  failed: 'failed',
  skipped: 'skipped',
  flood_wait: 'flood_wait',
} as const;

export interface JoinJob {
  id: number;
  accountId: number;
  /** @nullable */
  accountPhone?: string | null;
  linkId: number;
  /** @nullable */
  linkUrl?: string | null;
  status: JoinJobStatus;
  /** @nullable */
  errorCode?: string | null;
  /** @nullable */
  errorMessage?: string | null;
  createdAt: string;
}

export interface Collection {
  id: number;
  name: string;
  connectionString: string;
  dbName?: string;
  /** Field name in the collection containing the Telegram link */
  linkField: string;
  isActive: boolean;
  /** @nullable */
  lastSyncAt: string | null;
  syncedCount?: number;
  createdAt: string;
}

export interface CollectionInput {
  name: string;
  connectionString: string;
  dbName: string;
  linkField: string;
  isActive?: boolean;
}

export interface SyncResult {
  synced: number;
  duplicates: number;
  errors: number;
}

export interface BotStatus {
  running: boolean;
  /** @nullable */
  currentAccountId: number | null;
  /** @nullable */
  currentAccountPhone?: string | null;
  queueSize: number;
  /** Seconds between rotations based on account count */
  rotationInterval: number;
  /** @nullable */
  startedAt: string | null;
  totalJoinedToday?: number;
  totalFailedToday?: number;
}

export type ActivityEntryType = typeof ActivityEntryType[keyof typeof ActivityEntryType];


export const ActivityEntryType = {
  join_success: 'join_success',
  join_failed: 'join_failed',
  join_skipped: 'join_skipped',
  flood_wait: 'flood_wait',
  account_switched: 'account_switched',
  bot_started: 'bot_started',
  bot_stopped: 'bot_stopped',
  sync_completed: 'sync_completed',
} as const;

export interface ActivityEntry {
  id: number;
  type: ActivityEntryType;
  message: string;
  /** @nullable */
  accountPhone?: string | null;
  /** @nullable */
  linkUrl?: string | null;
  createdAt: string;
}

export type ListLinksParams = {
status?: ListLinksStatus;
source?: string;
};

export type ListLinksStatus = typeof ListLinksStatus[keyof typeof ListLinksStatus];


export const ListLinksStatus = {
  pending: 'pending',
  joined: 'joined',
  failed: 'failed',
  skipped: 'skipped',
} as const;

export type ListJobsParams = {
limit?: number;
/**
 * @nullable
 */
accountId?: number | null;
};