# lib/api-zod/src/generated/types/account.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Telegram Multi-Account Bot Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { AccountStatus } from './accountStatus';

export interface Account {
  id: number;
  phone: string;
  /** @nullable */
  label?: string | null;
  status: AccountStatus;
  joinedCount: number;
  failedCount: number;
  dailyLimit: number;
  /** Current delay in seconds before next join */
  currentDelay: number;
  /**
     * ISO timestamp when flood wait expires
     * @nullable
     */
  floodWaitUntil?: string | null;
  /** @nullable */
  sessionFile?: string | null;
  isPremium?: boolean;
  createdAt: string;
}