# lib/api-zod/src/generated/types/accountUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Telegram Multi-Account Bot Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { AccountUpdateStatus } from './accountUpdateStatus';

export interface AccountUpdate {
  label?: string;
  status?: AccountUpdateStatus;
  dailyLimit?: number;
  isPremium?: boolean;
}