# lib/api-zod/src/generated/types/activityEntry.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Telegram Multi-Account Bot Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { ActivityEntryType } from './activityEntryType';

export interface ActivityEntry {
  id: number;
  type: ActivityEntryType;
  message: string;
  /** @nullable */
  accountPhone?: string | null;
  /** @nullable */
  linkUrl?: string | null;
  createdAt: string;
}