# lib/api-zod/src/generated/types/bulkLinksInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Telegram Multi-Account Bot Manager API
 * OpenAPI spec version: 0.1.0
 */

export interface BulkLinksInput {
  urls: string[];
  source?: string;
}