# lib/api-zod/src/generated/types/collection.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Telegram Multi-Account Bot Manager API
 * OpenAPI spec version: 0.1.0
 */

export interface Collection {
  id: number;
  name: string;
  connectionString: string;
  dbName?: string;
  /** Field name in the collection containing the Telegram link */
  linkField: string;
  isActive: boolean;
  /** @nullable */
  lastSyncAt: string | null;
  syncedCount?: number;
  createdAt: string;
}